Create TRIBUTE Page
A tribute page is basically an overview of someone whom we admire in our life. In this article, we are creating a tribute webpage of Late A.P.J. Abdul Kalam Sir using HTML and CSS. We will add an image of him at the center (below the title) and create a box beneath that image.